test-symbols: Use readelf to find "_public_" symbols

+++ test/test-symbols.sh
@@ -2,10 +2,15 @@
 
 set -e
 
+# 77 skip test
+which readelf >/dev/null 2>&1 || exit 77
+test -e lib/libvarlink.sym || exit 77
+test -e libvarlink.a || exit 77
+
 rm -f symbols.list symbols.lib
 
-nm --dynamic --defined-only libvarlink.so.0 | awk '{ print $3 }' | sort > symbols.list
-grep -i VARLINK lib/libvarlink.sym | sed 's/ *;*{*//g' | sort > symbols.lib
+readelf -s -W libvarlink.a | grep 'FUNC    GLOBAL DEFAULT.*varlink_' | awk '{ print $8 }' | sort > symbols.list
+grep varlink_ lib/libvarlink.sym | sed 's/[ ;]//g' | sort > symbols.lib
 diff -u symbols.list symbols.lib
 r=$?
